# api/api.yaml
openapi: 3.1.0
info:
  title: Track Your Regions API
  version: 0.0.1
  contact:
    name: Uncovering World Support
    url: https://uncovering.world
    email: info@uncovering.world
  description: |
    The Track Your Regions API provides access to the region data used by the Uncovering World platform.
    It gives access to the regions (and related information) and to the other data that is stored in the platform.
    The API is currently in development and not yet available for public use.

servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.uncovering.world
    description: Production server, not yet available

tags:
  - name: regions
    description: Operations related to regions and divisions
  - name: hierarchy
    description: Operations related to hierarchies
  - name: geometry
    description: Operations related to region geometry
  - name: views
    description: Operations related to region grouping views

paths:

  # Region endpoints

  # Retrieve root regions (regions without a parent)
  /regions/root:
    get:
      operationId: getRootRegions
      tags: [regions]
      summary: Retrieve all root regions
      description: |
        Retrieve all root regions for a given hierarchy.
        If no hierarchy is specified, the default hierarchy is used.
      parameters:
        - name: hierarchyId
          in: query
          required: false
          description: |
            The hierarchy to retrieve root regions for.
            If not specified, the default hierarchy is used.
          schema:
            type: integer
      responses:
        '200':
          description: List of root regions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Region'
        '400':
          description: Bad request, most likely due to an invalid hierarchy ID

  # Retrieve information about a specific region (root or subregion)
  /regions/{regionId}:
    get:
      operationId: getRegion
      tags: [regions]
      summary: Retrieve information about a specific region
      description: |
        Retrieve information about a specific region, including its name, parent region, and whether it has subregions.
        If the region is a root region, the parent region is null.
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: integer
        - name: hierarchyId
          in: query
          required: false
          description: |
            The hierarchy to which the region belongs.
            If not specified, the default hierarchy is used.
          schema:
            type: integer
      responses:
        '200':
          description: Region information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Region'
        '400':
          description: Bad request
        '404':
          description: Region not found


  # Retrieve subregions for a specific region
  /regions/{regionId}/subregions:
    get:
      operationId: getSubregions
      tags: [regions]
      summary: Retrieve subregions for a specific region
      description: |
        Retrieve subregions for a specific region.
        If the region has no subregions, a 204 response is returned.
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: integer
        - name: getAll
          in: query
          required: false
          description: Whether to retrieve all nested subregions or just the immediate subregions.
          schema:
            type: boolean
        - name: hierarchyId
          in: query
          required: false
          description: |
            The hierarchy to which the region belongs.
            If not specified, the default hierarchy is used.
          schema:
            type: integer
      responses:
        '200':
          description: List of subregions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Region'
        '204':
          description: Region has no subregions
        '400':
          description: Bad request
        '404':
          description: Region not found

  # Retrieve the parent regions for a specific region
  /regions/{regionId}/ancestors:
    get:
      operationId: getRegionAncestors
      tags: [regions]
      summary: Retrieve the parent regions for a specific region
      description: |
        Retrieve the parent regions for a specific region, including the region itself.
        The regions are returned in order, with the root region first and the region itself last.
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: integer
        - name: hierarchyId
          in: query
          required: false
          description: |
            The hierarchy to which the region belongs.
            If not specified, the default hierarchy is used.
          schema:
            type: integer
      responses:
        '200':
          description: List of parent regions, including the region itself, in order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Region'
        '400':
          description: Bad request
        '404':
          description: Region not found

  /regions/{regionId}/geometry:
    get:
      operationId: getRegionGeometry
      tags: [regions, geometry]
      summary: Retrieve the geometry for a specific region
      description: |
        Retrieve the geometry for a specific region.
        If the region has no geometry, the resolution of empty geometries is controlled by the `resolveEmpty`
        query parameter. If `resolveEmpty` is true, the geometry of the region is computed by aggregating the
        geometries of the subregions. If `resolveEmpty` is false, a 204 response is returned.
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: integer
        - name: resolveEmpty
          in: query
          required: false
          description: Whether to resolve empty geometries by aggregating the geometries of the subregions.
          schema:
            type: boolean
        - name: hierarchyId
          in: query
          required: false
          description: |
            The hierarchy to which the region belongs.
            If not specified, the default hierarchy is used.
          schema:
            type: integer
      responses:
        '200':
          description: Geometry of the region, in GeoJSON format, always a MultiPolygon
          content:
            application/json:
              schema:
                type: object
                properties:
                  geometry:
                    $ref: '#/components/schemas/MultiPolygonGeometry'
        '204':
          description: Region has no geometry
        '400':
          description: Bad request
        '404':
          description: Region not found

  # List the available hierarchies
  /regions/hierarchies:
    get:
      operationId: getHierarchies
      tags: [regions, hierarchy]
      summary: List the available hierarchies
      description: |
        List the available hierarchies.
        The default hierarchy is always included in the list.
      responses:
        '200':
          description: List of available hierarchies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hierarchy'

  # Search for regions by name
  /regions/search:
    get:
      operationId: searchRegions
      tags: [regions]
      summary: Search for regions by name
      description: |
        Search for regions by name.
        The search is case-insensitive and matches partial words.
        The search is performed on the name of the region and all its ancestors.
        The results are ordered by the relevance of the match, with the most relevant match first.
      parameters:
        - name: query
          required: true
          in: query
          schema:
            type: string
        - name: hierarchyId
          in: query
          required: false
          description: |
            The hierarchy to search in.
            If not specified, the default hierarchy is used.
          schema:
            type: integer
      responses:
        '200':
          description: List of regions matching the query
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
        '204':
          description: No regions match the query
        '400':
          description: Bad request

  /regions/{regionId}/siblings:
    get:
      operationId: getRegionSiblings
      tags: [regions]
      summary: Retrieve the sibling regions for a specific region
      description: |
        Retrieve the sibling regions for a specific region, including the region itself.
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: integer
        - name: hierarchyId
          in: query
          required: false
          description: |
            The hierarchy to which the region belongs.
            If not specified, the default hierarchy is used.
          schema:
            type: integer
      responses:
        '200':
          description: List of sibling regions, including the region itself
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Region'
        '400':
          description: Bad request
        '404':
          description: Region not found

  # View endpoints

  # Retrieve all views
  /views:
    get:
      operationId: getViews
      tags: [views]
      summary: Retrieve all views
      description: |
        Retrieve all views for a given hierarchy.
        Views are groupings of regions based on specific criteria.
      parameters:
        - name: hierarchyId
          in: query
          required: false
          description: |
            The hierarchy to retrieve views for.
            If not specified, the default hierarchy is used.
          schema:
            type: integer
        - name: includeInactive
          in: query
          required: false
          description: |
            Whether to include inactive views in the results.
            Defaults to false.
          schema:
            type: boolean
      responses:
        '200':
          description: List of views
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/View'
        '204':
          description: No views found
        '400':
          description: Bad request

    post:
      operationId: createView
      tags: [views]
      summary: Create a new view
      description: |
        Create a new view for grouping regions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - hierarchyId
              properties:
                name:
                  type: string
                  description: Name of the view
                description:
                  type: string
                  description: Description of the view
                hierarchyId:
                  type: integer
                  description: The hierarchy this view belongs to
                isActive:
                  type: boolean
                  description: Whether the view is active
                  default: true
      responses:
        '201':
          description: View created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/View'
        '400':
          description: Bad request
        '404':
          description: Hierarchy not found
        '500':
          description: Internal server error

  # Retrieve a specific view
  /views/{viewId}:
    get:
      operationId: getViewById
      tags: [views]
      summary: Retrieve a specific view
      description: |
        Retrieve information about a specific view by its ID.
      parameters:
        - name: viewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: View information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/View'
        '404':
          description: View not found
        '500':
          description: Internal server error

    put:
      operationId: updateView
      tags: [views]
      summary: Update a view
      description: |
        Update an existing view's information.
      parameters:
        - name: viewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the view
                description:
                  type: string
                  description: Description of the view
                isActive:
                  type: boolean
                  description: Whether the view is active
      responses:
        '200':
          description: View updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/View'
        '400':
          description: Bad request
        '404':
          description: View not found
        '500':
          description: Internal server error

    delete:
      operationId: deleteView
      tags: [views]
      summary: Delete a view
      description: |
        Delete a view and all its region associations.
      parameters:
        - name: viewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: View deleted successfully
        '404':
          description: View not found
        '500':
          description: Internal server error

  # Get regions in a view
  /views/{viewId}/regions:
    get:
      operationId: getViewRegions
      tags: [views]
      summary: Get all regions in a view
      description: |
        Retrieve all regions that belong to a specific view.
      parameters:
        - name: viewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of regions in the view
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Region'
        '204':
          description: No regions found in this view
        '404':
          description: View not found
        '500':
          description: Internal server error

    post:
      operationId: addRegionsToView
      tags: [views]
      summary: Add regions to a view
      description: |
        Add one or more regions to a view.
      parameters:
        - name: viewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - regions
              properties:
                regions:
                  type: array
                  description: Array of regions to add to the view
                  items:
                    type: object
                    required:
                      - regionId
                    properties:
                      regionId:
                        type: integer
                        description: ID of the region to add
                      hierarchyId:
                        type: integer
                        description: Hierarchy ID (optional, uses view's hierarchy if not specified)
      responses:
        '201':
          description: Regions added to view successfully
        '400':
          description: Bad request
        '404':
          description: View or region not found
        '500':
          description: Internal server error

    delete:
      operationId: removeRegionsFromView
      tags: [views]
      summary: Remove regions from a view
      description: |
        Remove one or more regions from a view.
      parameters:
        - name: viewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - regions
              properties:
                regions:
                  type: array
                  description: Array of regions to remove from the view
                  items:
                    type: object
                    required:
                      - regionId
                    properties:
                      regionId:
                        type: integer
                        description: ID of the region to remove
                      hierarchyId:
                        type: integer
                        description: Hierarchy ID (optional, uses view's hierarchy if not specified)
      responses:
        '200':
          description: Regions removed from view successfully
        '400':
          description: Bad request
        '404':
          description: View not found
        '500':
          description: Internal server error

# Shared components

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  schemas:
    Region:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        parentRegionId:
          type: integer
        hasSubregions:
          type: boolean

    SearchResult:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        path:
          type: string
          description: A ' > ' delimited string representing the hierarchical path of the region.

    CRS:
      type: object
      properties:
        type:
          type: string
          enum: [ name ]
          description: Type of CRS; here, 'name' refers to a named CRS.
        properties:
          type: object
          properties:
            name:
              type: string
              description: Name of the CRS used; typically an EPSG code.
              example: "EPSG:4326"

    Coordinate:
      type: array
      minItems: 2
      maxItems: 2
      items:
        type: number
        format: double
      description: A longitude and latitude pair.

    LinearRing:
      type: array
      items:
        $ref: '#/components/schemas/Coordinate'
      description: A closed LineString, with four or more positions. The first and last positions are equivalent (they represent the same point).

    Polygon:
      type: array
      items:
        $ref: '#/components/schemas/LinearRing'
      description: An array of LinearRings where the first element is the outer boundary and any subsequent elements are inner boundaries (holes).

    MultiPolygonGeometry:
      type: object
      description: A GeoJSON object representing the geometry of the region as a MultiPolygon.
      properties:
        type:
          type: string
          enum: [ MultiPolygon ]
          description: Type of the geometry; for this API, always 'MultiPolygon'.
        coordinates:
          type: array
          items:
            $ref: '#/components/schemas/Polygon'
          description: An array of Polygons that make up the MultiPolygon.
        crs:
          $ref: '#/components/schemas/CRS'

    Hierarchy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

    View:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the view
        name:
          type: string
          description: Name of the view
        description:
          type: string
          description: Description of the view
        hierarchyId:
          type: integer
          description: The hierarchy this view belongs to
        isActive:
          type: boolean
          description: Whether the view is active
      required:
        - id
        - name
        - hierarchyId
        - isActive